check webgl in cosoulartcontainer component instead of parent
