+++ src/features/cosoul/MintPage.tsx
@@ -1,11 +1,10 @@
-import { useEffect, useRef, useState } from 'react';
+import { useState } from 'react';
 
 import { useLoginData } from 'features/auth';
 import { useQuery } from 'react-query';
 
 import isFeatureEnabled from 'config/features';
 // import { Button } from 'ui';
-import { Canvas } from 'ui';
 import { SingleColumnLayout } from 'ui/layouts';
 
 import { CoSoulArt } from './art/CoSoulArt';
@@ -36,33 +35,12 @@ export const MintPage = () => {
   const cosoul_data = query.data;
   const [minted, setMinted] = useState(false);
   const coSoulMinted = !!cosoul_data?.mintInfo;
-  const webglTest = useRef<HTMLCanvasElement>(null);
-  const [webglEnabled, setWebglEnabled] = useState(true);
-
-  useEffect(() => {
-    const canvas = webglTest.current;
-    const checkWebglEnabled = () => {
-      if (canvas) {
-        const webglEnabled = !!canvas.getContext('webgl2');
-        setWebglEnabled(webglEnabled);
-      }
-    };
-    checkWebglEnabled();
-  }, []);
 
   if (!isFeatureEnabled('cosoul')) {
     return <></>;
   }
   return (
     <>
-      <Canvas
-        ref={webglTest}
-        css={{
-          position: 'absolute',
-          zIndex: -1,
-          left: -5000,
-        }}
-      />
       {cosoul_data && (
         <SingleColumnLayout
           css={{
@@ -94,11 +72,7 @@ export const MintPage = () => {
                 onReveal={() => setMinted(true)}
               />
               <CoSoulComposition cosoul_data={cosoul_data} minted={minted}>
-                <CoSoulArtContainer
-                  cosoul_data={cosoul_data}
-                  minted={minted}
-                  webglEnabled={webglEnabled}
-                >
+                <CoSoulArtContainer cosoul_data={cosoul_data} minted={minted}>
                   <CoSoulArt pGive={cosoul_data.totalPgive} address={address} />
                 </CoSoulArtContainer>
               </CoSoulComposition>
